move utils tests
